requirements.txt

astroid==2.8.0
certifi==2021.5.30
cffi==1.15.0
charset-normalizer==2.0.6
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
defusedxml==0.7.1
Django==2.2.23
django-cors-headers==2.2.0
django-filter==2.4.0
django-haystack==3.1.0
django-js-asset==1.2.2
django-mptt==0.12.0
django-smtp-ssl==1.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
djangorestframework-jwt-proxy==0.1
djangorestframework-simplejwt==4.4.0
docopt==0.6.2
drf-haystack==1.8.11
drf-yasg==1.20.0
elasticsearch==7.12.1
elasticsearch-dsl==7.3.0
idna==3.2
inflection==0.5.1
isort==5.8.0
itypes==1.2.0
jieba==0.42.1
Jinja2==3.0.2
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
mccabe==0.6.1
mysqlclient==2.0.3
oauthlib==3.2.0
packaging==21.0
pipreqs==0.4.11
platformdirs==2.4.0
psycopg2==2.7.4
pycparser==2.21
PyJWT==1.7.0
pylint==2.11.1
pylint-django==2.4.4
pylint-plugin-utils==0.6
pyparsing==2.4.7
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.1
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
typed-ast==1.4.3
uritemplate==3.0.1
urllib3==1.26.7
Whoosh==2.7.4
wincertstore==0.2
wrapt==1.12.1
yarg==0.1.9
